+++ src/main/java/nl/martenm/servertutorialplus/ServerTutorialPlus.java
@@ -8,10 +8,10 @@ import nl.martenm.servertutorialplus.data.MySqlDataSource;
 import nl.martenm.servertutorialplus.events.*;
 import nl.martenm.servertutorialplus.helpers.Config;
 import nl.martenm.servertutorialplus.helpers.PluginUtils;
+import nl.martenm.servertutorialplus.hooks.PlaceholderAPIExpansion;
 import nl.martenm.servertutorialplus.language.Lang;
 import nl.martenm.servertutorialplus.managers.FlatFileManager;
 import nl.martenm.servertutorialplus.managers.NPCManager;
-import nl.martenm.servertutorialplus.managers.PlaceholderapiManager;
 import nl.martenm.servertutorialplus.managers.clickactions.ClickManager;
 import nl.martenm.servertutorialplus.objects.*;
 import nl.martenm.servertutorialplus.points.PointType;
@@ -125,7 +125,7 @@ public class ServerTutorialPlus extends JavaPlugin{
         if(Bukkit.getPluginManager().isPluginEnabled("PlaceholderAPI")){
             logger.info("PlaceholderAPI has been found!");
             placeholderAPI = true;
-            new PlaceholderapiManager(this).register();
+            new PlaceholderAPIExpansion(this).register();
         } else {
             logger.info("PlaceholderAPI was not found!");
             placeholderAPI = false;
